% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{colors}
\alias{colors}
\title{Color Scheme}
\format{
An object of class \code{list} of length 39.
}
\usage{
colors
}
\description{
Named list of blues, golds, and greys to be used as a color palette. These colors
are used as defaults for most internal plots.
}
\keyword{datasets}
