% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_axis.R
\name{date_axis}
\alias{date_axis}
\title{Date Axis}
\usage{
date_axis(
  date_breaks = "3 months",
  limits = NULL,
  labels = scales::label_date_short()
)
}
\arguments{
\item{date_breaks}{\code{str} Breaks to be used specified with the format '[#] [unit]', for
example : '2 weeks'. Default is '3 months'.}

\item{limits}{\verb{date vct} Axis limits to be used. Default is NULL, allowing
\code{ggplot2::scale_x_date()} to set the limits automatically.}

\item{labels}{\code{fct} Allows users to set their own labels if needed. Default is
\code{scales::label_date_short()}.}
}
\value{
\code{list} of \code{ggplot2} objects
}
\description{
Adds an attractive date axis to a timeseries plot. \code{date_axis} automatically removes
the title elemnt from the x-axis.
}
