% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_event.R
\name{geom_event}
\alias{geom_event}
\title{(Labeled) Event Line Geom}
\usage{
geom_event(
  line_position,
  label = "",
  label_position = NULL,
  orientation = "vertical",
  color = "black",
  linetype = "dashed",
  line_size = 0.5
)
}
\arguments{
\item{line_position}{\code{dbl / date / chr} Location on x or y axis where the line should be
anchored.}

\item{label}{\code{chr} Optional label to add above (if horizontal) or to the left of (if vertical)
the line. Default is no label.}

\item{label_position}{\code{dbl / date / chr} Location on the x (if horizontal) or y (if vertical)
axis to center the label. \strong{WARNING}, if a label is given to \code{geom_threshold} then \code{label_x}
\emph{must} be specified. Default is NULL.}

\item{orientation}{\code{chr} One of 'vertical' or 'horizontal' indicating the orientation of the
line. Default is 'vertical'.}

\item{color}{\code{chr} Color of line and text. Default is 'black'.}

\item{linetype}{\code{chr} Line type to be used; must be an appropriate \code{ggplot2} linetype. Default is
'dashed'.}

\item{line_size}{\code{dbl} Line thickness. Default is 0.5.}
}
\value{
\code{ggplot} geom
}
\description{
A geom to add a threshold line (with an optional label) to a ggplot.
}
