% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_time_periods.R
\name{geom_time_periods}
\alias{geom_time_periods}
\title{Label Sections of a Timeseries}
\usage{
geom_time_periods(y, periods, color = "#c5c5c5")
}
\arguments{
\item{y}{\code{dbl} Y axis value for section labels.}

\item{periods}{\code{df} Data frame contaning time period labels, start dates, and end dates. These
columns \emph{must} be named "label", "start", and "end" respectively.}

\item{color}{\code{str} Color to use for the vertical lines and labels. Default is grey (#c5c5c5).}
}
\value{
\code{list} of \code{ggplot2} objects
}
\description{
Adds vertical lines to a time series plot cutting the plot into different sections,
typically different waves of an epidemic. A horizontally centered label is added to each section
at a user specified height. \strong{Warning} : this is an opinionated function with specific column
name requirements in the \code{periods} argument.
}
