% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epicurve.R
\name{plot_epicurve}
\alias{plot_epicurve}
\title{Basic Epicurve}
\usage{
plot_epicurve(
  df,
  y = "cases",
  x = "date",
  color = "#2E4172",
  date_breaks = "2 months",
  date_limits = NULL,
  y_limits = NULL,
  y_breaks = 3,
  y_range = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  ...
)
}
\arguments{
\item{df}{\code{dataframe / tibble / datatable} Dataset.}

\item{y}{\code{chr} Character string of column to be plotted on the y axis.}

\item{x}{\code{chr} Column name of x axis data. Default is 'date'. \strong{Warning}: this column \emph{must} be
datetime formatted.}

\item{color}{\code{chr} Bar color to be used. Default is #2E4172 (dark blue).}

\item{date_breaks}{\code{chr} Character string indicating the time betweek x-axis ticks. Must be
expressed in the format '(#) (time unit)', for example '1 week', or '2 months'. \strong{Warning}:
this parameter must be all lowercase; '1 Week', for example, will return an error. Default is
'2 months'.}

\item{date_limits}{\verb{date vct} Axis limits to be used. Default is NULL, allowing
\code{ggplot2::scale_x_date()} to set the limits automatically.}

\item{y_limits}{\code{vct} Vector of two dbls indicating the lower and upper bounds of the y axis,
eg: \code{c(0, 2000)}. If NULL, limits are calculated automatically based on the data. Default is
NULL. \strong{NOTE}: providing \code{y_limits} will overide \code{y_breaks}.}

\item{y_breaks}{\code{int} Number of desired breaks on the y-axis; to be passed to \code{\link[base:pretty]{base::pretty()}}.
Default is 3.}

\item{y_range}{\verb{dbl series} Series to be used by \code{\link[base:pretty]{base::pretty()}} when calculating y-axis breaks.}

\item{xlab}{\code{chr} X-axis label.}

\item{ylab}{\code{chr} Y-axis label.}

\item{title}{\code{chr} Title.}

\item{subtitle}{\code{chr} Subtitle.}

\item{caption}{\code{chr} Caption.}

\item{...}{Parameters to be sent to theme_basic, eg \code{base_size}.}
}
\value{
\code{ggplot} of timeseries data.
}
\description{
Basic epicurve with the option to automatically aggregate the data.
}
