% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Static Choropleth}
\usage{
plot_map(
  df,
  x,
  x_range = NULL,
  date = NULL,
  map = NULL,
  factor_palette = NULL,
  color_palette = NULL,
  n_breaks = 5,
  na_color = "grey",
  border_color = "white",
  border_size = 0,
  percent = FALSE,
  si_notation = TRUE,
  keywidth = 70,
  label = FALSE,
  label_column = "label",
  label_size = 3,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  legend_title = NULL
)
}
\arguments{
\item{df}{\code{dataframe / tibble / datatable} Dataset.}

\item{x}{\code{chr} Column name of variable to be plotted in the choropleth.}

\item{x_range}{\code{vector} Min and max value to be used when setting the color scale of \code{x}. If NULL,
the min and max observed in \code{x} will be used. Default is NULL.}

\item{date}{\code{date} Date to be plotted. If NULL, the most recent available date is used. Default
is NULL.}

\item{map}{\code{sf} An optional simple features dataset containing GIS data to join with df if
not using one of the default maps. Default is NULL.}

\item{factor_palette}{\code{vector} Named list of colors (values) and label names (keys) to use for
the color scale. This option is intended for categorical variables and if present \code{plot_map()}
will ignore \code{n_breaks} and \code{color_palette}.}

\item{color_palette}{\code{palette} Color palette to be used. If none is provided, a reversed version
of 'YlOrRd' from \code{\link[RColorBrewer:ColorBrewer]{RColorBrewer::brewer.pal()}} will be used.}

\item{n_breaks}{\code{int} Number of desired breaks in \code{x}; to be passed to \code{\link[base:pretty]{base::pretty()}}. Default
is 5.}

\item{na_color}{\code{chr} Color to use for shapes with no data. Default is grey.}

\item{border_color}{\code{chr} Color of the borders. Default is white.}

\item{border_size}{\code{int} Size of the borders. Default is 0 (no borders).}

\item{percent}{\code{bool} Whether to format the the legend labels as percents. Default is FALSE.}

\item{si_notation}{\code{bool} Whether to format the legend labels with a limited si notation (k for
thousands, M for millions). Default is FALSE.}

\item{keywidth}{\code{int} Width of legend color bar. Note this is the length of the whole bar, not a
single segment. Default is 70.}

\item{label}{\code{bool} If TRUE, id labels will appear on the map. Default is FALSE.}

\item{label_column}{\code{chr} String indicating the column name to be used for labeling if \code{label} is
TRUE. Default is 'label'.}

\item{label_size}{\code{int} Text size of labels; ignored if \code{label} is FALSE. Default is 3.}

\item{title}{\code{chr} Title.}

\item{subtitle}{\code{chr} Subtitle.}

\item{caption}{\code{chr} Caption.}

\item{legend_title}{\code{chr} Legend Title.}
}
\value{
\code{ggplot} map.
}
\description{
Static choropleth map of France.
}
