% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid.R
\name{plot_pyramid}
\alias{plot_pyramid}
\title{Pyramid Plot}
\usage{
plot_pyramid(
  df,
  id,
  val_left,
  val_right,
  text_size = 5,
  center_width = 0.5,
  fill = NULL,
  fills_left = colors$blue,
  color_text_left = "black",
  fills_right = colors$blue,
  color_text_right = "black",
  limits_left = NULL,
  limits_right = limits_left,
  label_left = val_left,
  label_right = val_right,
  ...
)
}
\arguments{
\item{df}{\code{dataframe / tibble / datatable} Dataset.}

\item{id}{\code{chr} Column name of id data, to be shown in the middle of the pyramid.}

\item{val_left}{\code{chr} Column name of left-hand side data.}

\item{val_right}{\code{chr} Column name of right-hand side data.}

\item{text_size}{\code{int} Size of text labels in the middle. Default is 3.}

\item{center_width}{\code{flt} Relative width of the center section (with the text) compared to the
plot areas on either side. Default is 0.5 (50\% as big).}

\item{fill}{\code{chr} Column name of variable to be mapped to the bar fill. Default is NULL.}

\item{fills_left}{\code{chr/vct} Color(s) to use for the bar fill on the left-hand side. When \code{fill}
is specified, a vector of colors should be supplied. Default is colors$blue.}

\item{color_text_left}{\code{chr} Color to use for the axis text and ticks on the left-hand axis.
Default is 'black'.}

\item{fills_right}{\code{chr/vct} Color(s) to use for the bar fill on the right-hand side. When \code{fill}
is specified, a vector of colors should be supplied. Default is colors$blue.}

\item{color_text_right}{\code{chr} Color to use for the title and axis text / ticks on the right-hand axis.
Default is 'black'.}

\item{limits_left}{\code{vct} Limits to use for the left side of the pyramid. Default is to have
limits chosen automatically (NULL).}

\item{limits_right}{\code{vct} Limits to use for the right side of the pyramid. Default is to have
limits chosen automatically (NULL).}

\item{label_left}{\code{chr} Title to be shown for left-hand side (plotted on the top). Default is ''.}

\item{label_right}{\code{chr} Title to be shown for right-hand side (plotted on the top). Default is
''.}

\item{...}{Arguments to be passed to \code{\link[=theme_pyramid]{theme_pyramid()}}, eg: 'base_size'.}
}
\value{
\code{ggplot} Pyramid plot
}
\description{
Pyramid plot of two variables with a common id variable; eg: incidence and
positivity by department. Note, whereas traditional pyramid plots normally plot a single
indicator between groups for a single variable (for example counts of men and women for various
age bands), this pyramid is intended to be flexible enough to plot two unrelated indicators. As
a result, \code{plot_pyramid()} requres the data for the left-hand and right-hand sides to be in \emph{two
separate columns}.
}
\section{Functions}{
\itemize{
\item \code{plot_pyramid()}: Create pyramid plot

}}
