% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pyramid.R
\name{plot_pyramid_base}
\alias{plot_pyramid_base}
\title{Pyramid Plot Base}
\usage{
plot_pyramid_base(
  df,
  id,
  val,
  fill = NULL,
  fill_values = NULL,
  color_text = "black",
  axis_label = "",
  ...
)
}
\arguments{
\item{df}{\code{dataframe / tibble / datatable} Dataset.}

\item{id}{\code{chr} Column of id values to be set on vertical axis.}

\item{val}{\code{chr} Column of values to be plotted.}

\item{fill}{\code{chr} Column name of variable to be mapped to the bar fill. Default is NULL.}

\item{fill_values}{\code{chr/vct} Color(s) to use for the bar fill. When \code{fill} is specified, a vector
of colors should be supplied. Default is colors$blue.}

\item{color_text}{\code{chr} Color to use for axis text / ticks. Default is 'black'.}

\item{axis_label}{\code{chr} Optional title for axis values; to be displayed above the plot. Default
is ''.}

\item{...}{Arguments to be passed to \code{\link[=theme_pyramid]{theme_pyramid()}}, eg: 'base_size'.}
}
\value{
\code{ggplot} of base pyramid plot component.
}
\description{
Pyramid Plot Base
}
