% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_axis.R
\name{second_axis}
\alias{second_axis}
\title{Second Axis}
\usage{
second_axis(
  scale_factor,
  percent = FALSE,
  digits = 0,
  formatter = NULL,
  title = NULL,
  color = "#c5c5c5"
)
}
\arguments{
\item{scale_factor}{\code{dbl} Scale factor to convert main (left) axis into the new secondary (right)
axis}

\item{percent}{\code{bool} Convenience argument generating a percent formatted axis. Default is
\code{FALSE}.}

\item{digits}{\code{int} Number of decimal digits to use when \code{percent} is \code{TRUE}. Default is 0.}

\item{formatter}{\code{fct} Formatter to use for axis labels. This option is superceeded when
\code{percent} is \code{TRUE}.}

\item{title}{\code{str} Axis title. Default is NULL.}

\item{color}{\code{str} Axis color (ticks, title, values). Default is grey (#c5c5c5).}
}
\value{
\code{list} of \code{ggplot2} objects
}
\description{
Adds a secondary vertical axis on the right side of a ggplot based on a scale factor
of the main (left) vertical axis. *Note : * by  default the secondary axis will not have a line,
only ticks. If desired, an axis line must be added manually by the user.
}
