% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean ggplot Theme}
\usage{
theme_clean(
  base_size = 20,
  legend_x = NULL,
  legend_y = NULL,
  lines = FALSE,
  title_hjust = 0,
  subtitle_hjust = title_hjust
)
}
\arguments{
\item{base_size}{\code{int} Base size of the theme (for text sizing). Default is 14.}

\item{legend_x}{\code{dbl} Horizontal legend location. Default is NULL.}

\item{legend_y}{\code{dbl} Vertical legend location. Default is NULL.}

\item{lines}{\code{bool} Whether to plot y-axis grid lines. Default is FALSE.}

\item{title_hjust}{\code{dbl} Alignment of the plot title. Default is left aligned.}

\item{subtitle_hjust}{\code{dbl} Alignment of the subtitle. Default is to match the alignment
of the title.}
}
\value{
\code{ggplot} with base theme
}
\description{
Adds base theme to a \code{ggplot}; intended to be publication ready.
}
