% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_map}
\alias{theme_map}
\title{Add the basic map theme to a ggplot map}
\usage{
theme_map(base_size = 20, title_hjust = 0)
}
\arguments{
\item{base_size}{\code{int} Base size of the theme (for text sizing).}

\item{title_hjust}{\code{num} Hjust.}
}
\value{
\code{ggplot} with base map theme
}
\description{
Adds base map theme to a \code{ggplot}; intended to be publication ready.
}
