% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pyramid}
\alias{theme_pyramid}
\title{Add the basic pyramid theme to a ggplot}
\usage{
theme_pyramid(base_size = 20, color = "black")
}
\arguments{
\item{base_size}{\code{int} Base size of the theme (for text sizing).}

\item{color}{\code{chr} Color to be used for the horizontal axis. Default is 'black'.}
}
\value{
\code{ggplot} with base pyramid theme
}
\description{
Adds base pyramid theme to a \code{ggplot}; intended to be publication ready. Note, the
final pyramid plot is actually a grid of three plots, so this theme is to be used for each.
}
